project/requirements.txt
# # Web Framework
# fastapi
# uvicorn[standard]

# # Async HTTP Requests
# aiohttp

# # Database & ORM
# sqlalchemy
# aiosqlite
# asyncpg

# # Web Scraping & Parsing
# beautifulsoup4
# requests

# # Network Security Scanning
# python-nmap
# dnspython
# python-whois
# socketio
# whois

# # Cryptography & SSL Analysis
# cryptography
# pyOpenSSL

# # Background Tasks
# celery
# redis
# reportlab

# # Environment Variables Handling
# python-dotenv

# # Testing & Debugging
# pytest
# loguru

# python-socketio
# ssl

aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
bidict==0.23.1
billiard==4.2.1
bs4==0.0.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
orjson==3.10.15
packaging==24.2
pillow==11.1.0
# pip==25.0
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
Pygments==2.19.1
pyOpenSSL==25.0.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.3.4
python-multipart==0.0.20
python-nmap==0.7.1
python-socketio==5.7.2
python-whois==0.9.5
PyYAML==6.0.2
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
routers==0.10.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
# uvloop==0.21.0
validators==0.34.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
wsproto==1.2.0
yarl==1.18.3
flask
pdfplumber
matplotlib
